ntu method for q456, add week6

+++ ntu_arms.py
@@ -2,7 +2,7 @@
 import pieper as pp
 import craig as cg
 import numpy as np
-from sympy import Matrix, Symbol, Eq, solve, sin, cos
+from sympy import sin, cos
 
 def ntu_pieper():
     #warnings.filterwarnings('ignore')
@@ -12,17 +12,15 @@ def ntu_pieper():
                         [0, 340, 0],
                         [np.deg2rad(-90), -40, 338],
                         [np.deg2rad(90), 0, 0],
-                        [np.deg2rad(-90),0,0]])
+                        [np.deg2rad(-90),0,0]], dtype=float)
 
     cg.setDhTbl(dh_tbl)
-    #getcontext().prec=2
-    #getcontext().rounding = ROUND_UP
 
     ty = np.deg2rad(-60) # rotate y axis
     tcup_0_2s = np.array([[cos(ty), 0, sin(ty), 330], [0, 1, 0, 372],
-                            [-sin(ty), 0, cos(ty), 367], [0, 0, 0, 1]])
+                            [-sin(ty), 0, cos(ty), 367], [0, 0, 0, 1]], dtype=float)
     Tcup_6 = np.array([[0, 0, 1, 0], [0, -1, 0, 0], [1, 0, 0, 206],
-                        [0, 0, 0, 1]])
+                        [0, 0, 0, 1]], dtype=int)
 
     t6_0 = tcup_0_2s @ np.linalg.inv(Tcup_6)
     # t6_0=  np.round(t6_0.astype(np.double), 3)
